Add an ec2-specific executor
John's updates in the windows-test-harness, and his need to copy down the entire
Windows orb really irked me. I am not sure this implementation is any better, as
from a customer-facing perspective it might highlight our ec2 images.

https://github.com/circleci/windows-test-harness/pull/15/files#diff-1d37e48f9ceff6d8030570cd36286a61R7

I ran into a lot of frustration working with executors in writing this. In the
`default` executor I found myself with limited options if I wanted to be able
to customize the entire `image`. I could:
  1. Make a prefix for `version` that defaults to `windows-server-2019-vs2019:`
     For example, a `_family` value that could be submitted as ''
  2. Make a second executor that has a YOLO version
     Which is what I've done in this commit. It isn't ideal because
       - It duplicates a lot of logic
       - It is visible to the customer
       - Arguably, it makes the underworkings of our orb a lot more clear (less abstracted)

I would have liked to be able to:
  1. Have a hidden parameter (I don't want customers to know there's an image prefix)
  2. Pass a parameter set around -- I can copy this executor, but I don't want to
  3. Let an executor use an executor -- then I could make a `cci-test` executor (use a better name) that had an optional image field, and have the `default` executor use cci-test under the hood... I guess then we'd use cci-dev or something? seems like a lot of layers...
  4. Have a when/unless clause in the orb executor -- this only works in steps