Use a base executor for the Windows executors
Because now that we're having more than one executor,
things are getting kind-of messy.

Consideration: This looks like a worse customer experience,
so we're not actually sure we want to advocate for it.

At this point in time, we can probably kick the can down the
road until there are three executors... it'd be great if we
had some form of executor inheritance