DRY up the windows orb with YAML anchors
Now that we're starting to introduce multiple executors, quite a
bit of the YAML in the orb (especially descriptions and defaults)
is repeated. Given that the whole point of orbs is to allow better
duplication of code (well, one of the "whole point"s), we felt
that avoiding repeated language was a better direction to go in.

Co-authored by: alexey@circleci.com